+++ tests/Feature/SimpleFeatureTest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Tests\Unit;
+
+use PHPUnit\Framework\TestCase;
+use SebastianBergmann\RecursionContext\InvalidArgumentException;
+
+class SimpleFeatureTest extends TestCase
+{
+    /**
+     * @throws InvalidArgumentException
+     */
+    public function testTest(): void
+    {
+        $this->assertTrue(true);
+    }
+}